+++ src/constants/BASE_URL.ts
@@ -1,0 +1,1 @@
+export const api: string = import.meta.env.VITE_BASE_URL;
