--- src/services/DATABASE_URL.ts
@@ -1,1 +1,0 @@
-export const api: string | undefined = import.meta.env.VITE_BASE_URL;
