+++ src/template/Authentication/SignIn/index.tsx
@@ -2,40 +2,18 @@ import { Link } from "react-router-dom";
 import Banner from "../../../components/Banner";
 import Form from "../../../components/Form";
 import { SignInData } from "./data";
-import { useAuth } from "../../../hooks/useAuth";
-import { Login } from "../../../services/Login";
-import { toast } from "react-toastify";
+import { handleSubmit } from "../../../services/merchant/handleSubmit";
 
 export default function SignIn() {
-  const { auth, handleChange } = useAuth();
-  const data = SignInData(auth, handleChange);
+  const data = SignInData();
 
-  const handleSubmit2Fuctions = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    try {
-      const response = await Login(auth);
-      if (!response) {
-        toast.error("Resposta vazia");
-         return
-      }
-      if (!response.success) {
-        const firstError = Object.values(response.fields || {})[0];
-         toast.error(`${firstError}`);
-         return
-      }
-      const res = response;
-      console.log("Login: ", res);
-    } catch (error) {
-      console.error(error);
-    }
-  };
   return (
     <main className="w-full h-screen xl:h-screen px-6 py-6 flex lg:flex-row justify-center items-center gap-20">
       <div className="w-full h-fit flex flex-col gap-12 items-center">
         <Link to={"/"} className="font-semibold text-5xl">
           FOOD<span className="text-[#FDD835]">BRIDGE</span>
         </Link>
-        <Form {...data.formData} onSubmit={handleSubmit2Fuctions} />
+        <Form {...data.formData} onSubmit={handleSubmit} />
         <div className="w-full flex items-center gap-3">
           <div className="w-[70%] h-[1px] bg-black" />
           <p className="font-semibold text-[20px]">ou</p>
